refactor: rewrite fs with promises
