+++ src/fs/create.js
@@ -1,4 +1,4 @@
-import { writeFile, access, F_OK } from "fs";
+import { writeFile } from "fs/promises";
 import { join } from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
@@ -8,17 +8,11 @@ const __dirname = join(dirname(__filename), "files");
 const file = join(__dirname, "fresh.txt");
 
 const create = async () => {
-  access(file, F_OK, (e) => {
-    if (e) {
-      writeFile(
-        file,
-        "I am fresh and young",
-        (err) => err && console.error(err)
-      );
-    } else {
-      throw Error("FS operation failed");
-    }
-  });
+  try {
+    await writeFile(file, "I am fresh and young", { flag: "wx" });
+  } catch {
+    throw Error("FS operation failed");
+  }
 };
 
 await create();
