+++ src/fs/delete.js
@@ -1,4 +1,4 @@
-import { unlink } from "fs";
+import { unlink } from "fs/promises";
 import { join } from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
@@ -8,11 +8,11 @@ const folder = join(dirname(path), "files");
 const file = join(folder, "fileToRemove.txt");
 
 const remove = async () => {
-  unlink(file, (error) => {
-    if (error) {
-      throw Error("FS operation failed");
-    }
-  });
+  try {
+    await unlink(file);
+  } catch {
+    throw Error("FS operation failed");
+  }
 };
 
 await remove();
