+++ src/fs/list.js
@@ -1,4 +1,4 @@
-import { access, F_OK, readdir } from "fs";
+import { readdir } from "fs/promises";
 import { join } from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
@@ -7,16 +7,11 @@ const path = fileURLToPath(import.meta.url);
 const folder = join(dirname(path), "files");
 
 const list = async () => {
-  access(folder, F_OK, (e) => {
-    if (e) {
-      throw Error("FS operation failed");
-    } else {
-      readdir(folder, (error, files) => {
-        error && console.error(error);
-        console.log(files);
-      });
-    }
-  });
+  try {
+    console.log(await readdir(folder));
+  } catch {
+    throw Error("FS operation failed");
+  }
 };
 
 await list();
