+++ src/fs/read.js
@@ -1,4 +1,4 @@
-import { readFile } from "fs";
+import { readFile } from "fs/promises";
 import { join } from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
@@ -8,12 +8,11 @@ const folder = join(dirname(path), "files");
 const file = join(folder, "fileToRead.txt");
 
 const read = async () => {
-  readFile(file, "utf8", (error, content) => {
-    if (error) {
-      throw Error("FS operation failed");
-    }
-    console.log(content);
-  });
+  try {
+    console.log(await readFile(file, { encoding: "utf-8", flag: "r" }));
+  } catch {
+    throw Error("FS operation failed");
+  }
 };
 
 await read();
