+++ src/fs/rename.js
@@ -1,4 +1,4 @@
-import { access, F_OK, rename as renameMethod } from "fs";
+import { rename as renameMethod } from "fs/promises";
 import { join } from "path";
 import { fileURLToPath } from "url";
 import { dirname } from "path";
@@ -9,17 +9,11 @@ const file = join(folder, "wrongFilename.txt");
 const newFile = join(folder, "properFilename.md");
 
 const rename = async () => {
-  access(newFile, F_OK, (e) => {
-    if (e) {
-      renameMethod(file, newFile, (error) => {
-        if (error) {
-          throw Error("FS operation failed");
-        }
-      });
-    } else {
-      throw Error("FS operation failed");
-    }
-  });
+  try {
+    await renameMethod(file, newFile);
+  } catch {
+    throw Error("FS operation failed");
+  }
 };
 
 await rename();
